+++ CHANGELOG.md
@@ -1,4 +1,18 @@
-## 0.0.1
+<!--
+ * @Author: gaoyong06@qq.com 
+ * @Date: 2020-11-30 12:56:02 
+ * @Last Modified by: gaoyong06@qq.com
+ * @Last Modified time: 2020-11-30 12:56:02
+-->
+
+## 0.0.3
+
+  * Update documentation
+  * Bugfix：Android Compile fail
 
 ## 0.0.2
-    2020/12/02
+
+  * Add support for ios platform
+## 0.0.1
+
+  * Initial Version of the library